experimental and trunk: poppler 0.60.1
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@6481 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/libs/poppler/poppler-src/poppler/StructTreeRoot.cc
@@ -8,6 +8,7 @@
 // Copyright 2014 Fabio D'Urso <fabiodurso@hotmail.it>
 // Copyright 2017 Jan-Erik S <janerik234678@gmail.com>
 // Copyright 2017 Albert Astals Cid <aacid@kde.org>
+// Copyright 2017 Adrian Johnson <ajohnson@redneon.com>
 //
 //========================================================================
 
@@ -66,8 +67,8 @@ void StructTreeRoot::parse(Dict *root)
             error(errSyntaxError, -1, "Nums item at position {0:d} is wrong type ({1:s})", i, index.getTypeName());
             continue;
           }
-          const unsigned idx = index.getInt();
-          if (idx < 0 || idx >= parentTree.size()) {
+          const int idx = index.getInt();
+          if (idx < 0 || idx >= (int)parentTree.size()) {
             error(errSyntaxError, -1, "Nums item at position {0:d} is invalid value ({1:d}): [0..{2:d}]", i, idx, parentTree.size() - 1);
             continue;
           }
